requirements.txt
asttokens==2.4.1
azure-functions==1.20.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
ipykernel==6.29.4
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
pandas==2.2.2
parso==0.8.3
pip==24.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0
pytz==2024.1
pywin32==306
pyzmq==25.1.2
scikit-surprise==1.1.4
scipy==1.13.1
setuptools==70.1.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
wcwidth==0.2.13
wheel==0.43.0
